Refactored new classes out into new files
