+++ core/src/main/kotlin/injection/Context.kt
@@ -52,7 +52,7 @@ import story.StoryManager
 import ui.Hud
 import ui.IUserInterface
 import messaging.MessageHandler
-import turbofacts.FotW
+import turbofacts.NewFactsOfTheWorld
 import turbofacts.NewStoryManager
 
 object Context {
@@ -93,7 +93,7 @@ object Context {
             bindSingleton(RayHandler(inject(), 500, 500))
             bindSingleton(MessageHandler())
             bindSingleton(NewStoryManager())
-            bindSingleton(FotW { key -> inject<NewStoryManager>().apply { this.needsChecking = true } })
+            bindSingleton(NewFactsOfTheWorld { key -> inject<NewStoryManager>().apply { this.needsChecking = true } })
             bindSingleton(FactsOfTheWorld(
                 Gdx.app.getPreferences("TurboRocket"),
                 { key -> inject<MessageHandler>().sendMessage(Message.FactUpdated(key))}
